requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
apscheduler>=3.10
requests>=2.31
